requirements.txt
tensorflow>=2.10.0
opencv-python>=4.7.0
numpy>=1.24.0
Pillow>=9.5.0